requirements.txt
# Core Dependencies
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.9
python-dotenv==1.2.1

# Template Engine
Jinja2==3.1.4
MarkupSafe==2.1.5

# Data Processing
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.16.3
joblib==1.5.2

# Pydantic and Type Hints
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing-extensions==4.12.2

# Web and API
starlette==0.50.0
anyio==4.12.0
httptools==0.7.1
websockets==15.0.1
uvloop==0.22.1
watchfiles==1.1.1
h11==0.16.0

# Development Tools
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
PyYAML==6.0.3

# Jupyter (for model development)
notebook==7.2.0
ipykernel==7.1.0
ipython==9.7.0
matplotlib-inline==0.2.1
jupyter-client==8.6.3
jupyter-core==5.9.1

# Other Utilities
decorator==5.2.1
packaging==25.0
platformdirs==4.5.0
psutil==7.1.3
pyzmq==27.1.0
six==1.17.0

# Testing
pytest==8.2.0
pytest-cov==4.1.0
pytest-asyncio==0.23.7
httpx==0.27.0
